+++ assets/javascripts/kitten/hoc/with-tooltip.js
@@ -5,22 +5,18 @@ import { TooltipNew } from 'kitten/components/tooltips/tooltip-new'
 export const withTooltip = (WrappedComponent, wrappedProps) => {
   const WithTooltip = props => {
     const { children,
-            className,
-            tooltipClassName,
             id,
             ...others } = props
 
     return (
-      <div>
+      <div style={{ position: 'relative' }}>
         <TooltipNew id={ id } { ...others }>
           { children }
         </TooltipNew>
 
-        <div>
-          <WrappedComponent
-            data-for={ id }
-            aria-describedby={ id } />
-        </div>
+        <WrappedComponent
+          data-for={ id }
+          aria-describedby={ id } />
       </div>
     )
   }